In computer class

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,11 +8,11 @@ int main(void)
 {
 int i;
 int t1 = 1, t2 = 2;
-unsigned long nextTerm = t1 + t2;
+unsigned int  nextTerm = t1 + t2;
 	printf("1, 2, ");
 for (i = 3; i <= 50; ++i)
 {
-	printf("%lu, ", nextTerm);
+	printf("%u, ", nextTerm);
 	t1 = t2;
 	t2 = nextTerm;
 	nextTerm = t1 + t2;
